# backend_db/backend/openapi_extracted.yaml
openapi: 3.1.0
info:
  title: SelfDB
  description: Full Self-Hosted BaaS Built for AI Agents.
  version: 0.0.5
paths:
...
  /storage/files/:
    get:
      tags:
      - storage-files
      summary: List Files
      description: 'List files in a bucket with pagination, search, and sorting.


        - Public buckets: Accessible to anyone

        - Private buckets: Owner or admin only'
      operationId: list_files_storage_files__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: bucket_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          description: Bucket ID to list files from
          title: Bucket Id
        description: Bucket ID to list files from
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 100
          title: Page Size
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 100
            pattern: ^[\x20-\x7E]*$
          - type: 'null'
          title: Search
      - name: sort_by
        in: query
        required: false
        schema:
          enum:
          - created_at
          - updated_at
          - name
          - size
          type: string
          default: created_at
          title: Sort By
      - name: sort_order
        in: query
        required: false
        schema:
          enum:
          - asc
          - desc
          type: string
          default: desc
          title: Sort Order
      - name: X-API-Key
        in: header
        required: true
        schema:
          type: string
        description: API key required for all endpoints (validated by middleware)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDataResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '406':
          description: Not Acceptable (Missing API Key)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '413':
          description: File Too Large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /storage/files/{file_id}:
    get:
      tags:
      - storage-files
      summary: Get File Metadata
      description: Get file metadata by ID.
      operationId: get_file_metadata_storage_files__file_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: file_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: File Id
      - name: X-API-Key
        in: header
        required: true
        schema:
          type: string
        description: API key required for all endpoints (validated by middleware)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '406':
          description: Not Acceptable (Missing API Key)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '413':
          description: File Too Large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - storage-files
      summary: Delete File
      description: Delete a file. Only owner or admin can delete.
      operationId: delete_file_storage_files__file_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: file_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: File Id
      - name: X-API-Key
        in: header
        required: true
        schema:
          type: string
        description: API key required for all endpoints (validated by middleware)
      responses:
        '204':
          description: Successful Response
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '406':
          description: Not Acceptable (Missing API Key)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '413':
          description: File Too Large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - storage-files
      summary: Update File Metadata
      description: Update file metadata. Only owner or admin can update.
      operationId: update_file_metadata_storage_files__file_id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: file_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: File Id
      - name: X-API-Key
        in: header
        required: true
        schema:
          type: string
        description: API key required for all endpoints (validated by middleware)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Metadata
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '406':
          description: Not Acceptable (Missing API Key)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '413':
          description: File Too Large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoints__schema__ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    endpoints__schema__ErrorResponse:
      properties:
        detail:
          type: string
          title: Detail
      type: object
      required:
      - detail
      title: ErrorResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    FileResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        bucket_id:
          type: string
          format: uuid
          title: Bucket Id
        name:
          type: string
          title: Name
          examples:
          - document.pdf
          - photo.png
        path:
          type: string
          title: Path
          examples:
          - uploads/document.pdf
          - images/photo.png
        size:
          type: integer
          title: Size
          examples:
          - 1024
          - 2048576
        mime_type:
          type: string
          title: Mime Type
          examples:
          - application/pdf
          - image/png
        owner_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner Id
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
        checksum_sha256:
          anyOf:
          - type: string
          - type: 'null'
          title: Checksum Sha256
        version:
          type: integer
          title: Version
          default: 1
        is_latest:
          type: boolean
          title: Is Latest
          default: true
        deleted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deleted At
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - bucket_id
      - name
      - path
      - size
      - mime_type
      - created_at
      - updated_at
      title: FileResponse
      description: Response model for file (matches frontend FileItem interface).
...
    FileDataResponse:
      properties:
        files:
          items:
            $ref: '#/components/schemas/FileResponse'
          type: array
          title: Files
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
      type: object
      required:
      - files
      - total
      - page
      - page_size
      title: FileDataResponse
      description: Response model for file listing with metadata.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /users/token